requirements.txt
comodo-rest-api
cryptography
blinker
flask
flask-gssapi
Flask-Migrate
flask-restplus
flask-shell-ipython
gunicorn
mysql-connector
pyjsend
raven
